refactor: add auth service to collect authentication services helper
